Implement AuthMiddleware & RoomMiddleware

+++ back-end/routes/users.js
@@ -1,4 +1,4 @@
-const {Router } = require('express');
+const { Router } = require('express');
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
 const { PrismaClient } = require('@prisma/client');
@@ -7,19 +7,19 @@ const prisma = new PrismaClient();
 
 const router = Router();
 
-router.get('/', async (req, res) => {
+router.get('/', AuthMiddleware, async (req, res) => {
     try {
         const users = await prisma.user.findMany();
         res.json(users);
 
-    }   
+    }
     catch (error) {
         console.log(error);
         res.status(500).json({ error: 'Something went wrong' });
     }
 });
 
-router.get('/:userId', async (req, res) => {
+router.get('/:userId', AuthMiddleware, async (req, res) => {
     try {
         const user = await prisma.user.findUnique({ where: { id: parseInt(req.params.userId) } });
         res.json(user);
@@ -29,25 +29,10 @@ router.get('/:userId', async (req, res) => {
     }
 });
 
-router.post('/', async (req, res) => {
-    try {
-        const { email, password, firstname, lastname } = req.body;
-        const user = await prisma.user.create({
-            data: {
-                email,
-                password,
-                firstname,
-                lastname,
-            }
-        });
-        res.json(user);
-    } catch (error) {
-        console.log(error);
-        res.status(500).json({ error: 'Something went wrong' });
+router.put('/:userId', AuthMiddleware, async (req, res) => {
+    if (req.user.id !== parseInt(req.params.userId)) {
+        return res.status(403).json({ error: 'Unauthorized' });
     }
-});
-
-router.put('/:userId', async (req, res) => {
     try {
         const { email, password, firstname, lastname, role } = req.body;
         const user = await prisma.user.update({
@@ -67,7 +52,10 @@ router.put('/:userId', async (req, res) => {
     }
 });
 
-router.delete('/:userId', async (req, res) => {
+router.delete('/:userId', AuthMiddleware, async (req, res) => {
+    if (req.user.role !== 'ADMIN') {
+        return res.status(403).json({ error: 'Unauthorized' });
+    }
     try {
         const user = await prisma.user.delete({ where: { id: parseInt(req.params.userId) } });
         res.json(user);
